refactor: Modularize report generation and fix test type safety
- Extract helper functions from `report.py` to `report_sections.py`.
- Add type hints to all test functions.
- Fix type hints in `tests/test_fix_command.py` fixture.
