+++ tests/test_acl.py
@@ -6,7 +6,7 @@ from agent_scorecard.constants import PROFILES
 
 
 # TODO: Add type hints for Agent clarity
-def test_acl_calculation_logic():
+def test_acl_calculation_logic() -> None:
     """Tests the ACL calculation formula."""
     # Formula: ACL = CC + (LOC / 20)
 
@@ -29,7 +29,7 @@ def test_acl_calculation_logic():
     assert calculate_acl(cc, loc) == 20.0
 
 
-def test_scoring_with_acl_penalty(tmp_path: Path):
+def test_scoring_with_acl_penalty(tmp_path: Path) -> None:
     """Tests that a function with high ACL receives a penalty."""
 
     # RESOLUTION: We use the Advisor-Mode setup (Large Function) because
