+++ tests/test_analyzer_imports.py
@@ -4,7 +4,7 @@ from agent_scorecard.analyzer import get_import_graph, get_inbound_imports
 
 
 # TODO: Add type hints for Agent clarity
-def test_analyze_imports_internal_only(tmp_path):
+def test_analyze_imports_internal_only(tmp_path) -> None:
     # Create internal module
     """TODO: Add docstring for AI context."""
     (tmp_path / "internal.py").write_text("x = 1", encoding="utf-8")
