+++ tests/test_config.py
@@ -3,7 +3,7 @@ import tempfile
 from agent_scorecard.config import load_config, DEFAULT_CONFIG
 
 
-def test_load_config_defaults():
+def test_load_config_defaults() -> None:
     # Test loading from a directory with no pyproject.toml
     with tempfile.TemporaryDirectory() as tmpdir:
         config = load_config(tmpdir)
@@ -11,7 +11,7 @@ def test_load_config_defaults():
         assert config["thresholds"] == DEFAULT_CONFIG["thresholds"]
 
 
-def test_load_config_with_pyproject():
+def test_load_config_with_pyproject() -> None:
     with tempfile.TemporaryDirectory() as tmpdir:
         pyproject_content = """
 [tool.agent-scorecard]
@@ -32,7 +32,7 @@ type_safety = 80
         assert config["thresholds"]["type_safety"] == 80
 
 
-def test_load_config_invalid_toml():
+def test_load_config_invalid_toml() -> None:
     with tempfile.TemporaryDirectory() as tmpdir:
         with open(os.path.join(tmpdir, "pyproject.toml"), "w") as f:
             f.write("invalid = [")
